Rename energy button python package, scripts, and launcher
