+++ pyqt-apps/siriushla/as_ap_energybutton/energy_setter_window.py
@@ -1,16 +1,18 @@
 """Main window."""
 
 from qtpy.QtCore import Qt
-from qtpy.QtWidgets import QVBoxLayout, QWidget, QTabWidget
+from qtpy.QtWidgets import QTabWidget
 
 from siriushla.widgets import SiriusMainWindow
-from siriushla.energy_button.energy_setter import EnergySetter
-from siriushla.energy_button.energy_button import EnergyButton
+from siriushla.as_ap_energybutton.energy_setter import EnergySetter
+from siriushla.as_ap_energybutton.energy_button import EnergyButton
 
 
 class EnergySetterWindow(SiriusMainWindow):
+    """."""
 
     def __init__(self, parent=None):
+        """."""
         super().__init__(parent)
         self._widgets = list()
         self._setup_ui()
@@ -36,7 +38,7 @@ class EnergySetterWindow(SiriusMainWindow):
 if __name__ == '__main__':
     import sys
     from siriushla.sirius_application import SiriusApplication
-    
+
     app = SiriusApplication()
     w = EnergySetterWindow()
     w.show()
